feat: add styles to details element
